Add partitioned variant for parallel relationship scans

+++ community/kernel/src/main/java/org/neo4j/kernel/impl/newapi/DefaultRelationshipScanCursor.java
@@ -54,7 +54,7 @@ class DefaultRelationshipScanCursor extends DefaultRelationshipCursor implements
     boolean scanBatch(
             Read read,
             AllRelationshipsScan scan,
-            int sizeHint,
+            long sizeHint,
             LongIterator addedRelationships,
             boolean hasChanges,
             AccessMode accessMode) {
